Diverses modifications pour les tests (reverted from commit 635a6199bf9a77dabae267f0427f1ef3ac7bf4cd)

+++ backend/src/classification/DatasFIFO.java
@@ -2,17 +2,17 @@ package classification;
 
 public class DatasFIFO {
 	
-	double[][] tab;
+	float[][] tab;
 	int len;
 	
 	
 	// the size is 30 cause it matchs the duration of our recorded movements
 	public DatasFIFO() {
-		this.tab = new double[30][3];
+		this.tab = new float[30][3];
 		this.len=0;
 	}
 	
-	public void addData(double[] coordinates) {
+	public void addData(float[] coordinates) {
 		if (len < 29) {
 			tab[len+1]=coordinates;
 			len++;
